+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/boot/BootDef.java
@@ -1,11 +1,8 @@
-package uk.gov.hmcts.rse.ccd.lib.v2.data;
+package uk.gov.hmcts.rse.ccd.lib.boot;
 
-import com.microsoft.applicationinsights.TelemetryClient;
-import java.time.Clock;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
@@ -14,62 +11,51 @@ import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.annotation.PropertySources;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import uk.gov.hmcts.ccd.CoreCaseDataApplication;
-import uk.gov.hmcts.ccd.config.SwaggerConfiguration;
-import uk.gov.hmcts.ccd.data.AuthClientsConfiguration;
+import uk.gov.hmcts.ccd.definition.store.CaseDataAPIApplication;
+import uk.gov.hmcts.ccd.definition.store.SecurityConfiguration;
+import uk.gov.hmcts.ccd.definition.store.SwaggerConfiguration;
+import uk.gov.hmcts.ccd.definition.store.repository.AuthClientConfiguration;
+import uk.gov.hmcts.ccd.userprofile.endpoint.userprofile.UserProfileEndpoint;
 import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-import uk.gov.hmcts.reform.authorisation.validators.AuthTokenValidator;
-import uk.gov.hmcts.reform.authorisation.validators.ServiceAuthTokenValidator;
 import uk.gov.hmcts.reform.idam.client.IdamApi;
+import uk.gov.hmcts.rse.ccd.lib.YamlPropertySourceFactory;
 import uk.gov.hmcts.rse.ccd.lib.common.DBWaiter;
 
-@SpringBootConfiguration
-@EnableAutoConfiguration
 @ComponentScan(
     basePackageClasses = {
-        CoreCaseDataApplication.class,
+        CaseDataAPIApplication.class,
         DBWaiter.class
     },
-    excludeFilters = {
-        @ComponentScan.Filter(type= FilterType.REGEX, pattern = "uk\\.gov\\.hmcts\\.ccd\\.(definition|userprofile)\\..*"),
-        @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = {
-            uk.gov.hmcts.ccd.SecurityConfiguration.class,
-            CoreCaseDataApplication.class,
-            SwaggerConfiguration.class,
-            // We exclude this since it enables an unwanted S2S feign client
-            AuthClientsConfiguration.class
-        }),
-    }
-)
+    excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = {
+        CaseDataAPIApplication.class,
+        SwaggerConfiguration.class,
+        SecurityConfiguration.class,
+        AuthClientConfiguration.class
+    }))
 @PropertySources({
-    @PropertySource("classpath:datastore/application.properties"),
+    @PropertySource(value = {
+        "classpath:cftlib-defstore-es.yml"
+    }
+        , factory = YamlPropertySourceFactory.class),
+    @PropertySource("classpath:definitionstore/application.properties"),
+    @PropertySource("classpath:rse/definitionstore.properties"),
     @PropertySource("classpath:rse/application.properties"),
-    @PropertySource("classpath:rse/datastore.properties")
 })
 @EntityScan(basePackages = {
-    "uk.gov.hmcts.ccd.data"
+    "uk.gov.hmcts.ccd.definition.store"
 })
 @EnableJpaRepositories(basePackages = {
-    "uk.gov.hmcts.ccd.data"
+    "uk.gov.hmcts.ccd.definition.store"
 })
 @EnableFeignClients(
     clients = {
         IdamApi.class,
     })
-public class BootData {
-
-  @Bean
-  public Clock utcClock() {
-    return Clock.systemUTC();
-  }
-
-  @ConditionalOnMissingBean
-  @Bean
-  public TelemetryClient client() {
-    return new TelemetryClient();
-  }
+@SpringBootConfiguration
+@EnableAutoConfiguration
+public class BootDef {
 
   @Bean
   public AuthTokenGenerator authTokenGenerator(
@@ -78,10 +64,5 @@ public class BootData {
       final ServiceAuthorisationApi serviceAuthorisationApi
   ) {
     return AuthTokenGeneratorFactory.createDefaultGenerator(secret, microService, serviceAuthorisationApi);
-  }
-
-  @Bean
-  public AuthTokenValidator authTokenValidator(ServiceAuthorisationApi serviceAuthorisationApi) {
-    return new ServiceAuthTokenValidator(serviceAuthorisationApi);
   }
 }
