+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/boot/BootParent.java
@@ -1,15 +1,13 @@
-package uk.gov.hmcts.rse.ccd.lib.v2.lib;
+package uk.gov.hmcts.rse.ccd.lib.boot;
 
 import com.microsoft.applicationinsights.TelemetryClient;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 
-@ComponentScan(basePackageClasses = ParentContextConfiguration.class)
 @Configuration
-public class ParentContextConfiguration {
+public class BootParent {
 
   @ConditionalOnMissingBean
   @Bean
